PR feedback, add unit-tests
