+++ src/vs/workbench/contrib/chat/browser/actions/chatToolActions.ts
@@ -22,7 +22,8 @@ import { ConfirmedReason, IChatToolInvocation, ToolConfirmKind } from '../../com
 import { isResponseVM } from '../../common/model/chatViewModel.js';
 import { ChatModeKind } from '../../common/constants.js';
 import { IChatWidget, IChatWidgetService } from '../chat.js';
-import { isAgentHostBackendProvidedTool, ToolsScope } from '../widget/input/chatSelectedTools.js';
+import { ToolsScope } from '../widget/input/chatSelectedTools.js';
+import { isAgentHostBackendProvidedTool } from '../../common/tools/chatToolSelectionState.js';
 import { CHAT_CATEGORY } from './chatActions.js';
 import { showToolsPicker } from './chatToolPicker.js';
